requirement_text.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pandas>=1.5.3
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.2.0
torch>=2.0.0
transformers>=4.30.0
tqdm>=4.64.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.19.0
streamlit>=1.22.0
plotly>=5.13.0